# docs/swagger.yaml
definitions:
...
  github.com_VinCPR_backend_api.attendingResponse:
    properties:
      attending_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      mobile:
        type: string
    type: object
...
  github.com_VinCPR_backend_api.clinicalRotationEventDetailResponse:
    properties:
      attendings:
        items:
          $ref: '#/definitions/github.com_VinCPR_backend_api.attendingResponse'
        type: array
      end_date:
        type: string
      group_name:
        type: string
      hospital_name:
        type: string
      service_name:
        type: string
      specialty_name:
        type: string
      start_date:
        type: string
      students:
        items:
          $ref: '#/definitions/github.com_VinCPR_backend_api.studentResponse'
        type: array
    type: object
  github.com_VinCPR_backend_api.clinicalRotationEventResponse:
    properties:
      end_date:
        type: string
      event_id:
        type: integer
      group_name:
        type: string
      hospital_name:
        type: string
      service_name:
        type: string
      specialty_name:
        type: string
      start_date:
        type: string
    type: object
...
  github.com_VinCPR_backend_api.studentResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      mobile:
        type: string
      student_id:
        type: string
    required:
    - email
    type: object
...
info:
  contact: {}
paths:
...
  /rotation/attending:
    get:
      consumes:
      - application/json
      description: list calendar events for an attending in an academic year
      parameters:
      - description: academic year name
        in: query
        name: academicYearName
        required: true
        type: string
      - description: attending ID
        in: query
        name: attendingID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.clinicalRotationEventResponse'
            type: array
      summary: list calendar events for an attending in an academic year
      tags:
      - ClinicalRotation
...
  /rotation/detail:
    get:
      consumes:
      - application/json
      description: provide students and attending info and the clinical info
      parameters:
      - description: clinical rotation event ID
        in: query
        name: eventID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github.com_VinCPR_backend_api.clinicalRotationEventDetailResponse'
      summary: provide detail of a clinical rotation event
      tags:
      - ClinicalRotation
  /rotation/list/day:
    get:
      consumes:
      - application/json
      description: list clinical rotation event in that day
      parameters:
      - description: academic year name
        in: query
        name: academicYearName
        required: true
        type: string
      - description: day in format YYYY-MM-DD
        in: query
        name: day
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.clinicalRotationEventResponse'
            type: array
      summary: list clinical rotation event in that day
      tags:
      - ClinicalRotation
...
  /rotation/student:
    get:
      consumes:
      - application/json
      description: list calendar events for a student in an academic year
      parameters:
      - description: academic year name
        in: query
        name: academicYearName
        required: true
        type: string
      - description: student ID
        in: query
        name: studentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.clinicalRotationEventResponse'
            type: array
      summary: list calendar events for a student in an academic year
      tags:
      - ClinicalRotation
...
swagger: "2.0"